requirements.txt
web3>=6.0
eth-account>=0.11
anthropic>=0.40
curl_cffi>=0.7
rich>=13.0
requests>=2.31
python-dotenv>=1.0
fake-useragent>=1.4
httpx>=0.27
flask>=3.0
